+++ challenger/src/main/java/uk/co/compendiumdev/challenge/apimodel/ChallengeApiModel.java
@@ -1,5 +1,8 @@
 package uk.co.compendiumdev.challenge.apimodel;
 
+import static uk.co.compendiumdev.thingifier.apiconfig.EntityPatchUpdateStyle.JSON_MERGE_PATCH_RFC7396;
+import static uk.co.compendiumdev.thingifier.apiconfig.EntityPatchUpdateStyle.JSON_PATCH_RFC6902;
+import static uk.co.compendiumdev.thingifier.apiconfig.EntityPatchUpdateStyle.PARTIAL_JSON_UPDATE;
 import static uk.co.compendiumdev.thingifier.core.domain.definitions.field.definition.FieldType.STRING;
 
 import uk.co.compendiumdev.thingifier.Thingifier;
@@ -51,6 +54,10 @@ public class ChallengeApiModel {
         todoList.apiConfig().setApiToShowPrimaryKeyHeaderInResponse(false);
         todoList.apiConfig().statusCodes().setMaxRequestBodyLengthBytes(5000);
         todoList.apiConfig().setReturnSingleGetItemsAsCollection(true);
+        todoList.apiDefaults()
+                .writeMethods()
+                .entities()
+                .patchCan(PARTIAL_JSON_UPDATE, JSON_MERGE_PATCH_RFC7396, JSON_PATCH_RFC6902);
 
         return todoList;
     }